chore: move libs around and tests to own folder
