+++ lib/actor.js
@@ -1,5 +1,6 @@
 const { generateKey, encrypt, decrypt, addRecipient } = require('./crypto.js')
 const db = require('./db')
+const shortid = require('shortid')
 
 // TODO: all data should be in signed JWE/JWTs
 async function create(name) {
@@ -8,7 +9,7 @@ async function create(name) {
     key: await generateKey(name),
     createParcel: async () => {
       const parcel = {
-        id: Math.floor(Math.random()*99999999),
+        id: shortid.generate(),
         destination: await encrypt(actor.key, 'Ann Andersson, Långhalmsvägen 312, 510 88'),
         sender: `${name}, Avsändargatan 1, 123 45`,
         events: [],
